# services/disclosure_anchor/contracts/filing_api.openapi.yaml
components:
...
  schemas:
...
    ClassificationResponse:
      additionalProperties: false
      properties:
        class_map_version:
          title: Class Map Version
          type: string
        classes:
          items:
            $ref: '#/components/schemas/ProcessingClassV1'
          title: Classes
          type: array
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        processing_policy_available:
          title: Processing Policy Available
          type: boolean
        processing_policy_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Processing Policy Version
        rule_sets:
          items:
            $ref: '#/components/schemas/ClassificationRuleSetV1'
          title: Rule Sets
          type: array
      required:
      - class_map_version
      - processing_policy_version
      - processing_policy_available
      - classes
      - rule_sets
      title: ClassificationResponse
      type: object
    ClassificationRuleSetV1:
      additionalProperties: false
      properties:
        rule_count:
          title: Rule Count
          type: integer
        rule_set:
          title: Rule Set
          type: string
        version:
          title: Version
          type: string
      required:
      - rule_set
      - version
      - rule_count
      title: ClassificationRuleSetV1
      type: object
    DocumentListResponse:
      additionalProperties: false
      properties:
        items:
          items:
            $ref: '#/components/schemas/DocumentV1'
          title: Items
          type: array
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      required:
      - items
      - next_cursor
      title: DocumentListResponse
      type: object
...
    DocumentV1:
      additionalProperties: false
      properties:
        announcement_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Announcement Date
        company_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Ref
        content_categories:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Content Categories
        contract_version:
          title: Contract Version
          type: string
        correction_of_document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correction Of Document Id
        created_at:
          format: date-time
          title: Created At
          type: string
        current_processing_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Current Processing Run Id
        disclosure_topics:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Disclosure Topics
        document_id:
          title: Document Id
          type: string
        exchange:
          anyOf:
          - type: string
          - type: 'null'
          title: Exchange
        filing_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Filing Type
        market:
          anyOf:
          - type: string
          - type: 'null'
          title: Market
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        provider_document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider Document Id
        provider_metadata:
          additionalProperties: true
          title: Provider Metadata
          type: object
        publisher_categories:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Publisher Categories
        raw_file_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Raw File Hash
        report_period:
          anyOf:
          - type: string
          - type: 'null'
          title: Report Period
        security_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Security Code
        security_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Security Ref
        source_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Ref
        status:
          title: Status
          type: string
        superseded_by_document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Superseded By Document Id
        supersedes_document_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Supersedes Document Id
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - document_id
      - provider
      - provider_document_id
      - security_code
      - exchange
      - filing_type
      - disclosure_topics
      - title
      - announcement_date
      - report_period
      - raw_file_hash
      - status
      - current_processing_run_id
      - created_at
      - updated_at
      - contract_version
      - company_ref
      - security_ref
      - source_ref
      - supersedes_document_id
      - correction_of_document_id
      - superseded_by_document_id
      - provider_metadata
      - publisher_categories
      - market
      - content_categories
      title: DocumentV1
      type: object
...
    ProcessingClassV1:
      additionalProperties: false
      properties:
        disposition:
          title: Disposition
          type: string
        name:
          title: Name
          type: string
        priority:
          title: Priority
          type: integer
        zh:
          anyOf:
          - type: string
          - type: 'null'
          title: Zh
      required:
      - name
      - zh
      - priority
      - disposition
      title: ProcessingClassV1
      type: object
    ProcessingRunV1:
      additionalProperties: false
      properties:
        artifact_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Artifact Hash
        builder_rules_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Builder Rules Version
        content_hash_aggregate:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Hash Aggregate
        created_at:
          format: date-time
          title: Created At
          type: string
        document_id:
          title: Document Id
          type: string
        finished_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Finished At
        input_raw_file_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Input Raw File Hash
        is_active:
          title: Is Active
          type: boolean
        parser_backend:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser Backend
        parser_language:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser Language
        parser_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser Method
        parser_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser Name
        parser_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Parser Version
        processing_run_id:
          title: Processing Run Id
          type: string
        run_kind:
          title: Run Kind
          type: string
        started_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Started At
        status:
          title: Status
          type: string
        structure_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Structure Hash
        unit_build_attempt_count:
          title: Unit Build Attempt Count
          type: integer
        unit_build_status:
          title: Unit Build Status
          type: string
        unit_built_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Unit Built At
      required:
      - processing_run_id
      - document_id
      - run_kind
      - status
      - parser_name
      - parser_version
      - artifact_hash
      - content_hash_aggregate
      - structure_hash
      - is_active
      - started_at
      - finished_at
      - created_at
      - parser_backend
      - input_raw_file_hash
      - parser_method
      - parser_language
      - unit_build_status
      - unit_build_attempt_count
      - unit_built_at
      - builder_rules_version
      title: ProcessingRunV1
      type: object
...
info:
  title: disclosure_anchor
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/classification:
    get:
      operationId: get_classification_v1_classification_get
      parameters:
      - *id002
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema: *id001
          description: Contract version mismatch
        '422':
          content:
            application/json:
              schema: *id001
          description: Validation error
      summary: Get Classification
  /v1/documents:
    get:
      operationId: list_documents_v1_documents_get
      parameters:
      - in: query
        name: company_ref
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Ref
      - in: query
        name: security_code
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Security Code
      - in: query
        name: filing_type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Filing Type
      - in: query
        name: disclosure_topic
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Disclosure Topic
      - in: query
        name: report_period
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Report Period
      - in: query
        name: announcement_date_from
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Announcement Date From
      - in: query
        name: announcement_date_to
        required: false
        schema:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Announcement Date To
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: query
        name: cursor
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      - *id002
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema: *id001
          description: Contract version mismatch
        '422':
          content:
            application/json:
              schema: *id001
          description: Validation error
      summary: List Documents
  /v1/documents/{document_id}:
    get:
      operationId: get_document_v1_documents__document_id__get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      - in: query
        name: reject_superseded
        required: false
        schema:
          default: false
          title: Reject Superseded
          type: boolean
      - *id002
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentV1'
          description: Successful Response
        '400':
          content:
            application/json:
              schema: *id001
          description: Contract version mismatch
        '404':
          content:
            application/json:
              schema: *id001
          description: Not found
        '410':
          content:
            application/json:
              schema: *id001
          description: Gone superseded
        '422':
          content:
            application/json:
              schema: *id001
          description: Validation error
      summary: Get Document
  /v1/documents/{document_id}/runs:
    get:
      operationId: list_document_runs_v1_documents__document_id__runs_get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      - *id002
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProcessingRunV1'
                title: Response List Document Runs V1 Documents  Document Id  Runs
                  Get
                type: array
          description: Successful Response
        '400':
          content:
            application/json:
              schema: *id001
          description: Contract version mismatch
        '404':
          content:
            application/json:
              schema: *id001
          description: Not found
        '422':
          content:
            application/json:
              schema: *id001
          description: Validation error
      summary: List Document Runs
...